utils: add `stream` module for `CompilerOutputStream` and replace it everywhere
