+++ compiler/hash-driver/src/lib.rs
@@ -91,7 +91,7 @@ impl CompilerBuilder {
 pub mod utils {
     use std::io::Write;
 
-    use hash_pipeline::interface::CompilerOutputStream;
+    use hash_messaging::CompilerMessage;
     use hash_reporting::report::Report;
     use hash_utils::stream_writeln;
