+++ compiler/hash-pipeline/src/interface.rs
@@ -2,12 +2,11 @@
 //! that are used by the pipeline to run various stages that transform the
 //! provided sources into runnable/executable code.
 
-use std::sync::{Arc, Mutex};
-
 use hash_ast::node_map::NodeMap;
 use hash_reporting::report::Report;
 use hash_source::SourceId;
 pub use hash_utils::profiling::StageMetrics;
+use hash_utils::stream::CompilerOutputStream;
 
 use crate::{
     settings::{CompilerSettings, CompilerStageKind},
@@ -52,53 +51,6 @@ pub trait CompilerStage<StageCtx> {
     /// This function is used to to return the `stage` kind of
     /// this [CompilerStage].
     fn kind(&self) -> CompilerStageKind;
-}
-
-/// A [CompilerOutputStream] is used to specify where the output of the compiler
-/// should be written to. This is used by the [CompilerInterface] to provide
-/// the pipeline with the necessary information to write to the correct stream.
-#[derive(Debug)]
-pub enum CompilerOutputStream {
-    /// A [CompilerOutputStream] that points to the `stdout` stream.
-    Stdout(std::io::Stdout),
-
-    /// A [CompilerOutputStream] that points to the `stderr` stream.
-    Stderr(std::io::Stderr),
-
-    /// A [CompilerOutputStream] that is backed by a [Mutex] and a [Vec].
-    Owned(Arc<Mutex<Vec<u8>>>),
-}
-
-impl Clone for CompilerOutputStream {
-    fn clone(&self) -> Self {
-        match self {
-            CompilerOutputStream::Stdout(_) => CompilerOutputStream::Stdout(std::io::stdout()),
-            CompilerOutputStream::Stderr(_) => CompilerOutputStream::Stderr(std::io::stderr()),
-            CompilerOutputStream::Owned(stream) => CompilerOutputStream::Owned(stream.clone()),
-        }
-    }
-}
-
-impl std::io::Write for CompilerOutputStream {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        match self {
-            CompilerOutputStream::Stdout(stream) => stream.write(buf),
-            CompilerOutputStream::Stderr(stream) => stream.write(buf),
-            CompilerOutputStream::Owned(stream) => {
-                let mut stream = stream.lock().unwrap();
-                stream.extend_from_slice(buf);
-                Ok(buf.len())
-            }
-        }
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        match self {
-            CompilerOutputStream::Stdout(stream) => stream.flush(),
-            CompilerOutputStream::Stderr(stream) => stream.flush(),
-            CompilerOutputStream::Owned(_) => Ok(()),
-        }
-    }
 }
 
 /// The [CompilerInterface] serves as an interface between the created compiler
