+++ compiler/hash-utils/src/lib.rs
@@ -22,6 +22,7 @@ pub mod range_map;
 pub mod scoping;
 pub mod stack;
 pub mod state;
+pub mod stream;
 pub mod temp_writer;
 pub mod tree_writing;
